+++ examples/basevm/cmd/base-cli/cmd/spam.go
@@ -6,41 +6,17 @@ package cmd
 
 import (
 	"context"
-	"math/rand"
-	"os"
-	"os/signal"
-	"strings"
-	"sync"
-	"sync/atomic"
-	"syscall"
-	"time"
 
-	"github.com/ava-labs/avalanchego/utils/math"
+	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/hypersdk/chain"
 	"github.com/ava-labs/hypersdk/crypto"
 	"github.com/ava-labs/hypersdk/examples/basevm/actions"
 	"github.com/ava-labs/hypersdk/examples/basevm/auth"
-	"github.com/ava-labs/hypersdk/examples/basevm/consts"
-	trpc "github.com/ava-labs/hypersdk/examples/basevm/rpc"
-	"github.com/ava-labs/hypersdk/examples/basevm/utils"
+	brpc "github.com/ava-labs/hypersdk/examples/basevm/rpc"
 	"github.com/ava-labs/hypersdk/rpc"
 	hutils "github.com/ava-labs/hypersdk/utils"
 	"github.com/spf13/cobra"
-	"golang.org/x/sync/errgroup"
 )
-
-const (
-	feePerTx     = 2000
-	defaultRange = 32
-)
-
-type txIssuer struct {
-	c  *rpc.JSONRPCClient
-	tc *trpc.JSONRPCClient
-	d  *rpc.WebSocketClient
-
-	l              sync.Mutex
-	outstandingTxs int
-}
 
 var spamCmd = &cobra.Command{
 	Use: "spam",
@@ -49,407 +25,49 @@ var spamCmd = &cobra.Command{
 	},
 }
 
-func getRandomRecipient(self int, keys []crypto.PrivateKey) (crypto.PublicKey, error) {
-	if randomRecipient {
-		priv, err := crypto.GeneratePrivateKey()
-		if err != nil {
-			return crypto.EmptyPublicKey, err
-		}
-		return priv.PublicKey(), nil
-	}
-
-	// Select item from array
-	index := rand.Int() % len(keys)
-	if index == self {
-		index++
-		if index == len(keys) {
-			index = 0
-		}
-	}
-	return keys[index].PublicKey(), nil
-}
-
-func getRandomIssuer(issuers []*txIssuer) *txIssuer {
-	index := rand.Int() % len(issuers)
-	return issuers[index]
-}
-
 var runSpamCmd = &cobra.Command{
 	Use: "run",
 	RunE: func(*cobra.Command, []string) error {
-		ctx := context.Background()
-
-		// Select chain
-		chainID, uris, err := promptChain("select chainID")
-		if err != nil {
-			return err
-		}
-
-		// Select root key
-		keys, err := GetKeys()
-		if err != nil {
-			return err
-		}
-		if len(keys) == 0 {
-			return ErrNoKeys
-		}
-		cli := rpc.NewJSONRPCClient(uris[0])
-		networkID, _, _, err := cli.Network(ctx)
-		if err != nil {
-			return err
-		}
-		tcli := trpc.NewJSONRPCClient(uris[0], networkID, chainID)
-		hutils.Outf("{{cyan}}stored keys:{{/}} %d\n", len(keys))
-		balances := make([]uint64, len(keys))
-		for i := 0; i < len(keys); i++ {
-			address := utils.Address(keys[i].PublicKey())
-			balance, err := tcli.Balance(ctx, address)
-			if err != nil {
-				return err
-			}
-			balances[i] = balance
-			hutils.Outf(
-				"%d) {{cyan}}address:{{/}} %s {{cyan}}balance:{{/}} %s %s\n",
-				i,
-				address,
-				hutils.FormatBalance(balance),
-				consts.Symbol,
-			)
-		}
-		keyIndex, err := promptChoice("select root key", len(keys))
-		if err != nil {
-			return err
-		}
-		key := keys[keyIndex]
-		balance := balances[keyIndex]
-		factory := auth.NewED25519Factory(key)
-
-		// Distribute funds
-		numAccounts, err := promptInt("number of accounts")
-		if err != nil {
-			return err
-		}
-		numTxsPerAccount, err := promptInt("number of transactions per account per second")
-		if err != nil {
-			return err
-		}
-		witholding := uint64(feePerTx * numAccounts)
-		distAmount := (balance - witholding) / uint64(numAccounts)
-		hutils.Outf(
-			"{{yellow}}distributing funds to each account:{{/}} %s %s\n",
-			hutils.FormatBalance(distAmount),
-			consts.Symbol,
+		var bclient *brpc.JSONRPCClient
+		return handler.Root().Spam(maxTxBacklog, randomRecipient,
+			func(uri string, networkID uint32, chainID ids.ID) {
+				bclient = brpc.NewJSONRPCClient(uri, networkID, chainID)
+			},
+			func(pk crypto.PrivateKey) chain.AuthFactory {
+				return auth.NewED25519Factory(pk)
+			},
+			func(choice int, address string) (uint64, error) {
+				balance, err := bclient.Balance(context.TODO(), address)
+				if err != nil {
+					return 0, err
+				}
+				hutils.Outf(
+					"%d) {{cyan}}address:{{/}} %s {{cyan}}balance:{{/}} %s %s\n",
+					choice,
+					address,
+					handler.Root().ValueString(ids.Empty, balance),
+					handler.Root().AssetString(ids.Empty),
+				)
+				return balance, err
+			},
+			func(ctx context.Context, chainID ids.ID) (chain.Parser, error) {
+				return bclient.Parser(ctx)
+			},
+			func(pk crypto.PublicKey, amount uint64) chain.Action {
+				return &actions.Transfer{
+					To:    pk,
+					Value: amount,
+				}
+			},
+			func(cli *rpc.JSONRPCClient, pk crypto.PrivateKey) func(context.Context, uint64) error {
+				return func(ictx context.Context, count uint64) error {
+					_, _, err := sendAndWait(ictx, nil, &actions.Transfer{
+						To:    pk.PublicKey(),
+						Value: count, // prevent duplicate txs
+					}, cli, bclient, auth.NewED25519Factory(pk), false)
+					return err
+				}
+			},
 		)
-		accounts := make([]crypto.PrivateKey, numAccounts)
-		dcli, err := rpc.NewWebSocketClient(uris[0])
-		if err != nil {
-			return err
-		}
-		funds := map[crypto.PublicKey]uint64{}
-		parser, err := tcli.Parser(ctx)
-		if err != nil {
-			return err
-		}
-		var fundsL sync.Mutex
-		for i := 0; i < numAccounts; i++ {
-			// Create account
-			pk, err := crypto.GeneratePrivateKey()
-			if err != nil {
-				return err
-			}
-			accounts[i] = pk
-
-			// Send funds
-			_, tx, _, err := cli.GenerateTransaction(ctx, parser, nil, &actions.Transfer{
-				To:    pk.PublicKey(),
-				Value: distAmount,
-			}, factory)
-			if err != nil {
-				return err
-			}
-			if err := dcli.RegisterTx(tx); err != nil {
-				return err
-			}
-			funds[pk.PublicKey()] = distAmount
-
-			// Ensure Snowman++ is activated
-			if i < 10 {
-				time.Sleep(500 * time.Millisecond)
-			}
-		}
-		for i := 0; i < numAccounts; i++ {
-			_, dErr, result, err := dcli.ListenTx(ctx)
-			if err != nil {
-				return err
-			}
-			if dErr != nil {
-				return dErr
-			}
-			if !result.Success {
-				// Should never happen
-				return ErrTxFailed
-			}
-		}
-		hutils.Outf("{{yellow}}distributed funds to %d accounts{{/}}\n", numAccounts)
-
-		// Kickoff txs
-		clients := make([]*txIssuer, len(uris))
-		for i := 0; i < len(uris); i++ {
-			cli := rpc.NewJSONRPCClient(uris[i])
-			networkID, _, _, err := cli.Network(ctx)
-			if err != nil {
-				return err
-			}
-			tcli := trpc.NewJSONRPCClient(uris[i], networkID, chainID)
-			dcli, err := rpc.NewWebSocketClient(uris[i])
-			if err != nil {
-				return err
-			}
-			clients[i] = &txIssuer{c: cli, tc: tcli, d: dcli}
-		}
-		signals := make(chan os.Signal, 2)
-		signal.Notify(signals, syscall.SIGINT, syscall.SIGTERM)
-		var (
-			transferFee uint64
-			wg          sync.WaitGroup
-
-			l            sync.Mutex
-			confirmedTxs uint64
-			totalTxs     uint64
-		)
-
-		// confirm txs (track failure rate)
-		cctx, cancel := context.WithCancel(ctx)
-		defer cancel()
-		var inflight atomic.Int64
-		var sent atomic.Int64
-		var exiting sync.Once
-		for i := 0; i < len(clients); i++ {
-			issuer := clients[i]
-			wg.Add(1)
-			go func() {
-				for {
-					_, dErr, result, err := issuer.d.ListenTx(context.TODO())
-					if err != nil {
-						return
-					}
-					inflight.Add(-1)
-					issuer.l.Lock()
-					issuer.outstandingTxs--
-					issuer.l.Unlock()
-					l.Lock()
-					if result != nil {
-						if result.Success {
-							confirmedTxs++
-						} else {
-							hutils.Outf("{{orange}}on-chain tx failure:{{/}} %s %t\n", string(result.Output), result.Success)
-						}
-					} else {
-						// We can't error match here because we receive it over the wire.
-						if !strings.Contains(dErr.Error(), rpc.ErrExpired.Error()) {
-							hutils.Outf("{{orange}}pre-execute tx failure:{{/}} %v\n", dErr)
-						}
-					}
-					totalTxs++
-					l.Unlock()
-				}
-			}()
-			go func() {
-				<-cctx.Done()
-				for {
-					issuer.l.Lock()
-					outstanding := issuer.outstandingTxs
-					issuer.l.Unlock()
-					if outstanding == 0 {
-						_ = issuer.d.Close()
-						wg.Done()
-						return
-					}
-					time.Sleep(500 * time.Millisecond)
-				}
-			}()
-		}
-
-		// log stats
-		t := time.NewTicker(1 * time.Second) // ensure no duplicates created
-		defer t.Stop()
-		var psent int64
-		go func() {
-			for {
-				select {
-				case <-t.C:
-					current := sent.Load()
-					l.Lock()
-					if totalTxs > 0 {
-						hutils.Outf(
-							"{{yellow}}txs seen:{{/}} %d {{yellow}}success rate:{{/}} %.2f%% {{yellow}}inflight:{{/}} %d {{yellow}}issued/s:{{/}} %d\n", //nolint:lll
-							totalTxs,
-							float64(confirmedTxs)/float64(totalTxs)*100,
-							inflight.Load(),
-							current-psent,
-						)
-					}
-					l.Unlock()
-					psent = current
-				case <-cctx.Done():
-					return
-				}
-			}
-		}()
-
-		// broadcast txs
-		unitPrice, err := clients[0].c.SuggestedRawFee(ctx)
-		if err != nil {
-			return err
-		}
-		g, gctx := errgroup.WithContext(ctx)
-		for ri := 0; ri < numAccounts; ri++ {
-			i := ri
-			g.Go(func() error {
-				t := time.NewTimer(0) // ensure no duplicates created
-				defer t.Stop()
-
-				issuer := getRandomIssuer(clients)
-				factory := auth.NewED25519Factory(accounts[i])
-				fundsL.Lock()
-				balance := funds[accounts[i].PublicKey()]
-				fundsL.Unlock()
-				defer func() {
-					fundsL.Lock()
-					funds[accounts[i].PublicKey()] = balance
-					fundsL.Unlock()
-				}()
-				for {
-					select {
-					case <-t.C:
-						// Ensure we aren't too backlogged
-						if inflight.Load() > int64(maxTxBacklog) {
-							t.Reset(1 * time.Second)
-							continue
-						}
-
-						// Send transaction
-						start := time.Now()
-						selected := map[crypto.PublicKey]int{}
-						for k := 0; k < numTxsPerAccount; k++ {
-							recipient, err := getRandomRecipient(i, accounts)
-							if err != nil {
-								return err
-							}
-							v := selected[recipient] + 1
-							selected[recipient] = v
-							_, tx, fees, err := issuer.c.GenerateTransactionManual(parser, nil, &actions.Transfer{
-								To:    recipient,
-								Value: uint64(v), // ensure txs are unique
-							}, factory, unitPrice)
-							if err != nil {
-								hutils.Outf("{{orange}}failed to generate:{{/}} %v\n", err)
-								continue
-							}
-							transferFee = fees
-							if err := issuer.d.RegisterTx(tx); err != nil {
-								continue
-							}
-							balance -= (fees + uint64(v))
-							issuer.l.Lock()
-							issuer.outstandingTxs++
-							issuer.l.Unlock()
-							inflight.Add(1)
-							sent.Add(1)
-						}
-
-						// Determine how long to sleep
-						dur := time.Since(start)
-						sleep := math.Max(1000-dur.Milliseconds(), 0)
-						t.Reset(time.Duration(sleep) * time.Millisecond)
-					case <-gctx.Done():
-						return gctx.Err()
-					case <-cctx.Done():
-						return nil
-					case <-signals:
-						exiting.Do(func() {
-							hutils.Outf("{{yellow}}exiting broadcast loop{{/}}\n")
-							cancel()
-						})
-						return nil
-					}
-				}
-			})
-		}
-		if err := g.Wait(); err != nil {
-			return err
-		}
-
-		// Wait for all issuers to finish
-		hutils.Outf("{{yellow}}waiting for issuers to return{{/}}\n")
-		dctx, cancel := context.WithCancel(ctx)
-		go func() {
-			// Send a dummy transaction if shutdown is taking too long (listeners are
-			// expired on accept if dropped)
-			t := time.NewTicker(15 * time.Second)
-			defer t.Stop()
-			for {
-				select {
-				case <-t.C:
-					hutils.Outf("{{yellow}}remaining:{{/}} %d\n", inflight.Load())
-					_ = submitDummy(dctx, cli, tcli, key.PublicKey(), factory)
-				case <-dctx.Done():
-					return
-				}
-			}
-		}()
-		wg.Wait()
-		cancel()
-
-		// Return funds
-		hutils.Outf("{{yellow}}returning funds to %s{{/}}\n", utils.Address(key.PublicKey()))
-		var (
-			returnedBalance uint64
-			returnsSent     int
-		)
-		for i := 0; i < numAccounts; i++ {
-			balance := funds[accounts[i].PublicKey()]
-			if transferFee > balance {
-				continue
-			}
-			returnsSent++
-			// Send funds
-			returnAmt := balance - transferFee
-			_, tx, _, err := cli.GenerateTransaction(ctx, parser, nil, &actions.Transfer{
-				To:    key.PublicKey(),
-				Value: returnAmt,
-			}, auth.NewED25519Factory(accounts[i]))
-			if err != nil {
-				return err
-			}
-			if err := dcli.RegisterTx(tx); err != nil {
-				return err
-			}
-			returnedBalance += returnAmt
-
-			// Ensure Snowman++ is activated
-			if i < 10 {
-				time.Sleep(500 * time.Millisecond)
-			}
-		}
-		for i := 0; i < returnsSent; i++ {
-			_, dErr, result, err := dcli.ListenTx(ctx)
-			if err != nil {
-				return err
-			}
-			if dErr != nil {
-				return dErr
-			}
-			if !result.Success {
-				// Should never happen
-				return ErrTxFailed
-			}
-		}
-		hutils.Outf(
-			"{{yellow}}returned funds:{{/}} %s %s\n",
-			hutils.FormatBalance(returnedBalance),
-			consts.Symbol,
-		)
-		return nil
 	},
 }
