--- examples/tokenvm/cmd/token-cli/cmd/utils.go
@@ -1,37 +1,0 @@
-package cmd
-
-import (
-	"context"
-
-	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/vms/platformvm/warp"
-	"github.com/ava-labs/hypersdk/chain"
-	trpc "github.com/ava-labs/hypersdk/examples/tokenvm/rpc"
-	"github.com/ava-labs/hypersdk/rpc"
-)
-
-// TODO: use websockets
-func sendAndWait(
-	ctx context.Context, warpMsg *warp.Message, action chain.Action, cli *rpc.JSONRPCClient,
-	tcli *trpc.JSONRPCClient, factory chain.AuthFactory, printStatus bool,
-) (bool, ids.ID, error) {
-	parser, err := tcli.Parser(ctx)
-	if err != nil {
-		return false, ids.Empty, err
-	}
-	submit, tx, _, err := cli.GenerateTransaction(ctx, parser, warpMsg, action, factory)
-	if err != nil {
-		return false, ids.Empty, err
-	}
-	if err := submit(ctx); err != nil {
-		return false, ids.Empty, err
-	}
-	success, err := tcli.WaitForTransaction(ctx, tx.ID())
-	if err != nil {
-		return false, ids.Empty, err
-	}
-	if printStatus {
-		handler.Root().PrintStatus(tx.ID(), success)
-	}
-	return success, tx.ID(), nil
-}
